video can now be uploaded with product

+++ admin/product-edit.php
@@ -4,6 +4,11 @@
 if(isset($_POST['form1'])) {
 	$valid = 1;
 
+	if(empty($_POST['brand_id'])) {
+        $valid = 0;
+        $error_message .= "You must have to select a brand<br>";
+    }
+
     if(empty($_POST['tcat_id'])) {
         $valid = 0;
         $error_message .= "You must have to select a top level category<br>";
@@ -36,6 +41,19 @@ if(isset($_POST['form1'])) {
 
     $path = $_FILES['p_featured_photo']['name'];
     $path_tmp = $_FILES['p_featured_photo']['tmp_name'];
+
+    $path2 = $_FILES['p_featured_video']['name'];
+    $path_tmp2 = $_FILES['p_featured_video']['tmp_name'];
+
+    if($path2!='') {
+        $ext2 = pathinfo( $path2, PATHINFO_EXTENSION );
+        $file_name2 = basename( $path2, '.' . $ext2 );
+        if( $ext2 !='mp4' ) {
+            $valid = 0;
+            $error_message .= 'You must have to upload mp4 video file<br>';
+        }
+       $video_name = 'featured-video'.mt_rand().'.'.$ext2;
+    }
 
     if($path!='') {
         $ext = pathinfo( $path, PATHINFO_EXTENSION );
@@ -89,7 +107,7 @@ if(isset($_POST['form1'])) {
             }            
         }
 
-        if($path == '') {
+        if($path == '' && $path2 == '') {
         	$statement = $pdo->prepare("UPDATE tbl_product SET 
         							p_name=?, 
         							p_old_price=?, 
@@ -102,7 +120,8 @@ if(isset($_POST['form1'])) {
         							p_return_policy=?,
         							p_is_featured=?,
         							p_is_active=?,
-        							ecat_id=?
+        							ecat_id=?,
+        							brand_id=?
 
         							WHERE p_id=?");
         	$statement->execute(array(
@@ -118,22 +137,37 @@ if(isset($_POST['form1'])) {
         							$_POST['p_is_featured'],
         							$_POST['p_is_active'],
         							$_POST['ecat_id'],
+        							$_POST['brand_id'],
         							$_REQUEST['id']
         						));
         } else {
-
-        	unlink('../assets/uploads/'.$_POST['current_photo']);
+        	if ($path !='') {
+        		unlink('../assets/uploads/'.$_POST['current_photo']);
 
 			$final_name = 'product-featured-'.mt_rand().$_REQUEST['id'].'.'.$ext;
         	move_uploaded_file( $path_tmp, '../assets/uploads/'.$final_name );
-
+        	$statement = $pdo->prepare("UPDATE tbl_product SET 
+        							p_featured_photo=?
+        							WHERE p_id=?");
+        	$statement->execute(array($final_name,$_REQUEST['id']));
+        	}
+        	
+        	if ($path2 !='') {
+        		if(isset($_POST['current_video'])){
+        			unlink('../assets/video/'.$_POST['current_video']);
+        		}
+        		move_uploaded_file( $path_tmp2, '../assets/video/'.$video_name );
+        		$statement = $pdo->prepare("UPDATE tbl_product SET 
+        							p_featured_video=?
+        							WHERE p_id=?");
+        	$statement->execute(array($video_name,$_REQUEST['id']));
+        	}
 
         	$statement = $pdo->prepare("UPDATE tbl_product SET 
         							p_name=?, 
         							p_old_price=?, 
         							p_current_price=?, 
         							p_qty=?,
-        							p_featured_photo=?,
         							p_description=?,
         							p_short_description=?,
         							p_feature=?,
@@ -141,7 +175,8 @@ if(isset($_POST['form1'])) {
         							p_return_policy=?,
         							p_is_featured=?,
         							p_is_active=?,
-        							ecat_id=?
+        							ecat_id=?,
+        							brand_id=?
 
         							WHERE p_id=?");
         	$statement->execute(array(
@@ -149,7 +184,6 @@ if(isset($_POST['form1'])) {
         							$_POST['p_old_price'],
         							$_POST['p_current_price'],
         							$_POST['p_qty'],
-        							$final_name,
         							$_POST['p_description'],
         							$_POST['p_short_description'],
         							$_POST['p_feature'],
@@ -158,8 +192,11 @@ if(isset($_POST['form1'])) {
         							$_POST['p_is_featured'],
         							$_POST['p_is_active'],
         							$_POST['ecat_id'],
+        							$_POST['brand_id'],
         							$_REQUEST['id']
         						));
+
+        	
         }
 		
 
@@ -232,6 +269,7 @@ foreach ($result as $row) {
 	$p_current_price = $row['p_current_price'];
 	$p_qty = $row['p_qty'];
 	$p_featured_photo = $row['p_featured_photo'];
+	$p_featured_video = $row['p_featured_video'];
 	$p_description = $row['p_description'];
 	$p_short_description = $row['p_short_description'];
 	$p_feature = $row['p_feature'];
@@ -240,6 +278,7 @@ foreach ($result as $row) {
 	$p_is_featured = $row['p_is_featured'];
 	$p_is_active = $row['p_is_active'];
 	$ecat_id = $row['ecat_id'];
+	$brand_id = $row['brand_id'];
 }
 
 $statement = $pdo->prepare("SELECT * 
@@ -298,6 +337,24 @@ foreach ($result as $row) {
 
 				<div class="box box-info">
 					<div class="box-body">
+						<div class="form-group">
+							<label for="" class="col-sm-3 control-label">Brand Name <span>*</span></label>
+							<div class="col-sm-4">
+								<select name="brand_id" class="form-control select2">
+		                            <option value="">Select Brand</option>
+		                            <?php
+		                            $statement = $pdo->prepare("SELECT * FROM tbl_brand ORDER BY brand_name ASC");
+		                            $statement->execute();
+		                            $result = $statement->fetchAll(PDO::FETCH_ASSOC);   
+		                            foreach ($result as $row) {
+		                                ?>
+		                                <option value="<?php echo $row['brand_id']; ?>" <?php if($row['brand_id'] == $brand_id){echo 'selected';} ?>><?php echo $row['brand_name']; ?></option>
+		                                <?php
+		                            }
+		                            ?>
+		                        </select>
+							</div>
+						</div>
 						<div class="form-group">
 							<label for="" class="col-sm-3 control-label">Top Level Category Name <span>*</span></label>
 							<div class="col-sm-4">
@@ -359,13 +416,13 @@ foreach ($result as $row) {
 							</div>
 						</div>	
 						<div class="form-group">
-							<label for="" class="col-sm-3 control-label">Old Price<br><span style="font-size:10px;font-weight:normal;">(In USD)</span></label>
+							<label for="" class="col-sm-3 control-label">Old Price<br><span style="font-size:10px;font-weight:normal;"></span></label>
 							<div class="col-sm-4">
 								<input type="text" name="p_old_price" class="form-control" value="<?php echo $p_old_price; ?>">
 							</div>
 						</div>	
 						<div class="form-group">
-							<label for="" class="col-sm-3 control-label">Current Price <span>*</span><br><span style="font-size:10px;font-weight:normal;">(In USD)</span></label>
+							<label for="" class="col-sm-3 control-label">Current Price <span>*</span><br><span style="font-size:10px;font-weight:normal;"></span></label>
 							<div class="col-sm-4">
 								<input type="text" name="p_current_price" class="form-control" value="<?php echo $p_current_price; ?>">
 							</div>
@@ -437,6 +494,23 @@ foreach ($result as $row) {
 							<label for="" class="col-sm-3 control-label">Change Featured Photo </label>
 							<div class="col-sm-4" style="padding-top:4px;">
 								<input type="file" name="p_featured_photo">
+							</div>
+						</div>
+						<?php if($p_featured_video !=''): ?>
+						<div class="form-group">
+							<label for="" class="col-sm-3 control-label">Existing Featured video</label>
+							<div class="col-sm-4" style="padding-top:4px;">
+								<video  controls style="width:150px;">
+								  <source src="../assets/video/<?php echo $p_featured_video; ?>" type="video/mp4" >
+								</video>
+								<input type="hidden" name="current_video" value="<?php echo $p_featured_video; ?>">
+							</div>
+						</div>
+						<?php endif; ?>
+						<div class="form-group">
+							<label for="" class="col-sm-3 control-label">Add / Change Featured Video </label>
+							<div class="col-sm-4" style="padding-top:4px;">
+								<input type="file" name="p_featured_video">
 							</div>
 						</div>
 						<div class="form-group">
